Use city id from the database

+++ api/src/commonMain/kotlin/fobo66/valiutchik/api/ApiDataSource.kt
@@ -34,7 +34,7 @@ interface ApiDataSource {
      */
     suspend fun loadExchangeRates(
         currencies: List<String>,
-        cityIndex: Int
+        cityIndex: Long
     ): List<CurrencyRateSource>
 
     /**
